refactor: split stats export helpers
